kbd: remove chvt service
The chvt service is being moved to the installer/puppet, since we need to
make the foreground container configurable. The current recipe based
approach is not easy to configure, and forces all containers to contain
the service.

To make a container active/foreground after boot, the "active" parameter
can be added to the vty attribute.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>