cube: container lifecycle management
We hadd the ability to start/stop and list containers, but no simple way
to add and remove containers from the system.

This commit provides the capability to dom0 to add and remove containers
from the system.

It also arranges for the cube-cmd-server to restart when containers are
added, since that is required to establish communication channels to
those new containers.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>