mutt: avoid host contamination when looking for maildir
The configure script will look in hard coded locations for the
maildir, all of which are locations on the host. Rather then allow
this to happen we will pass the expected maildir.

This also fixes a build issue if none of the hardcoded paths exist on
the builder. Preventing the build error:

    configure: error: "Could not determine where new mail is stored."

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>