cube-console: allows nsenter from non dom0 containers
When managing multiple containers from cube-desktop or cube-server, having
a shortcut to enter the namespace of those containers is often useful.

With this commit, it is possible to shortcut a namespace enter from a
non dom0 container (the same way that shotcut consoles work).

This is an "expert" command, and will remain undocumented

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>