rndmac: A systemd service to randomize the MAC addr for hardware
In some cases the MAC address will need to be randomized for
development hardware and SDKs.  This new service will never be on by
default.  It is activated by passing the kernel command line
parameters:

   systemd.wants=rndmac.service systemd.setenv=rndmac=eth0

This service will not impact the startup time of the target or affect
systemd because it is configured to off by default. It has a very
minimal foot print addition to the file system of around 9K.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>