Revert "distro: ensure that no extra locales are built"
This reverts commit 295d16c9c92db7d8ce743d01c1508a4399e52e80.

The UTF8 locale are now a hard requirement in order to build
packages like py3, and if we cant revector LC_ALL off of "C"
and onto a suitable UTF8, we'll get lots of warnings.

In addition, the build system will probably have a hard stop
test added to it that checks for this early on sometime in
the near future, so if there are still runtime mismatch issues,
we'll have to deal with them another way if they reappear.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>