screen: Add -noexit and -nocheckpid options
The -noexit allows screen to not exit when the last window closes.
This allows for re-instantiation of a pre-defined shell when using
screen as a console service with a restricted screenrc.

The -nocheckpid option allows screen to be used across containers when
the socket pipe is visible to both containers.  The screen server and
client would check each other's pid and refuse to communicate.  The
pid name spaces are different between containers and this check must
be conditionally avoided.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>