screen-getty: Allow screen to shim between a serial port and agetty
This patch provides 2 programs, systemd service and a restricted screenrc:
   /sbin/agetty-shell
       - A tiny c program to run agetty as a screen shell
   /sbin/screen-getty
       - Shim to handle tty_open and start screen running agetty-shell
   /lib/systemd/system/screen-getty@.serivce
       - Compatible service to serial-getty@.service but uses screen
   /etc/default/screen-gettyrc
       - Restricted command screenrc

The idea is to replace agetty with screen running directly on a serial
port.  The screen client will not exit and it will loop just like
agetty starting from systemd.  Additional (up to 8) agetty login
sessions may be spawned by using "^W C".

Another motivation behind this patch was to provide the ability to
share the console of a container.  The capability to share the console
of the container is controlled in the containers autodev script.  The
autodev script can spawn a new screen session by invoking something
like:

   screen -S console -x -X screen lxc-console -n CONTAINER -t 0

The net result of using screen in this manager allows it to mimic the
behavor of a VT console where you would press control-alt-FX where FX
is a F1, F2, F3 etc... to change consoles which are mapped to
container domains.  In the case of the screen enabled serial port the
end user presses "Control-W 0", "Control-W 1" etc...

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>