cube-ctl: check if a container is known before attempting to stop it
Currently if you issue the 'cube-ctl stop' command and provide an
invalid container name you will get a failure exit code but no
indication as to why the command failed. This is inconsistent with
'cube-ctl del' where we do check the validity of the container name
first.

Put the code used by 'cube-ctl del' to check the validity of the
container name into a function called check_is_known_container() and
use this for 'del' as well as 'stop' to provide the user with useful
feedback.

NOTE: the logic used in 'del' has changed some but as far as I can
tell the updated logic I have used here is more complete. ie. It
properly handles the case where 'i_am_dom0' is false.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>