cube-ctl: provide user with more info when 'stop' fails
Unfortunately the logging found in cube-ctl (and dom0-contctl) is
inconsistent. For example if you attempt to start an already running
container you will get a message indicating as such, but attempting to
stop an already stopped container will return a failure code, but not
print any error info. The messaging is also inconsistently formatted,
com0-contctl will prefix with 'Error' whereas cube-ctl will prefix
with '[Error]'. Since there is ongoing work to replace the lxc backend
I am not going to attempt to correct all the messaging in this
commit. I will however make the messaging between 'start' and 'stop'
operations in cube-ctl more consistent.

If we attempt to stop a container and we get a failure return code
check if the container is STOPPED. If it is stopped let the user know
that they are possibly attempting to stop an already stopped
container. Ideally the check would be done before performing the
operation, but since these operations are slow we only check
after. This means we have to be careful with the error message wording
as we can't be 100% certain that the container was previously stopped.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>