dtach: Add additional functions for cross container commands
In order to implement the full cube-console and command stream handler
previously implemented with socat a few more extensions were needed to
dtach.

"-Q" argument: The cross container command console infrastructure
requires the ability to terminate the server if the client detaches.

"-R" argument: The output stream is raw when on the client side such
that the \r characters will appear if redirecting the output to a
file.  In order to avoid an additional fork and pipe to clean up the
output stream add the ability to the client to strip the \r characters
if needed.

"-T" argument: When executing commands from systemd in one container
to systemd in another container there may not be a controlling
terminal.  This argument allows dtach to run without a controlling
terminal for the purpose of collecting output from command pipes and
processing it elsewhere.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>