cube-cmd: Allow running cube-cmd without a controlling terminal
By default dtach checks for a controlling terminal with tcgetattr()
and will exit if it doesn't find one.  The "-T" argument to dtach
allows dtach to operate without a controlling terminal and provide the
unmodifed raw output stream.  If dtach is used with "-T" and it finds
a controlling terminal, it will still make use of it and properly put
the controlling terminal into raw mode.

The cube-cmd should use dtach with the -T argument else you will get a
failure occasionally with backgrounded commands.  The failure can be
observed with the following test case:

root@cube-dom0:~# cube-cmd lxc-info -n cube-desktop < /dev/null
/usr/bin/dtach: Attaching to a session requires a terminal.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>