images: allow user to overwrite IMAGE_FSTYPES
There is no reason to strictly enforce the use of one IMAGE_FSTYPE and
this is something which is often customized in local.conf (we even
include this in our local.conf.sample), so change this to an early
assignment allowing a user to set this how they wish in their
local.conf or elsewhere.

Since the machine definition or other configs may attempt to enable
the 'live' IMAGE_FSTYPE and this brings in busybox via
VIRTUAL-RUNTIME_base-utils and VIRTUAL-RUNTIME_base-utils-hwclock,
which we currently don't implement a workaround for, we explicitly
remove the 'live' IMAGE_FSTYPE to prevent its use.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>