lxc: Add lxc.uncontain = 1
A new option is added to the lxc configuration called "lxc.uncontain =
1".  This option will control if the CLONE_UNCONTAIN flag is passed to
the clone() system call when setting up the name space.

The CLONE_UNCONTAIN will be used to setup a "super user" name space
which could make use of other programs to re-enter the top level name
space in order to create new processes or additional name spaces.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>