overc-conftools: redefine the meaning of NETWORK_DEVICE and NETWORK_DEVICE_CLASSES
In the previous framework, NETWORK_DEVICE represents the network interfaces
that needed to passed through to the network prime and the interfaces
that needed to be managed by systemd-networkd. The NETWORK_DEVICE_CLASSES
was used only when NETWORK_DEVICE was set to "all", and it reprents the network
interfaces that needed to passed through to the network prime. This mechanism
mix the two variables meaning up.

This commit will make NETWORK_DEVICE only reprents the network interfaces
that needed to be managed by systemd-networkd and the NETWORK_DEVICE_CLASSES
was used to represent the network interfaces that needed to passed through to
the network prime.

If a system didn't use networkd to manage the network devices, please don't
set NETWORK_DEVICE or set it empty; If a system want to passed through
some network devices to the network prime, please set the network devices
string pattern to NETWORK_DEVICE_CLASSES, such as "wl* eth* en*".

Signed-off-by: fli <fupan.li@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>