core: remove samba from overc-common pkgdefs
This was added in an earlier attempt at comparing other default distro pkglists
to ours.  However samba isn't actually installed by default on most distros
and we don't strictly need it in the base layer here either.

Since it has been problematic getting it built recently, and since it has a
significant number of dependencies, we remove it from common and let the
real users re-add it at a higher level.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>