meta-overc: upgrade/rollback/factory-reset grub-efi (when it exists)
The kernel and grub-efi related files are on a separate part of /boot. In the
existing implementation of system upgrade/rollback/factory-reset, we only
copu the kernel image into the boot section when the system prepares for
upgrade/rollback/factory-reset. This meant that grub-efi related files were
not copied into the boot section. If grub-efi related files are
upgraded/changed during a system upgrade, then we should also upgrade/reback
the those files also.

Since the grub-efi related files are the key section of secure boot, that's
why we should also take care of them once they are upgraded.

Signed-off-by: fli <fupan.li@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>