distro: update overc SANITY_TESTED_DISTROS to match upstream.
We copied the data for overc a while ago ; I looked into sharing
it vs. copy/fork but that seems to fail the risk vs. reward ratio
at this point in time.  So just update our copy.

We keep our overc entry for the self-builder, and I add ubu 16.10
since I've been personally using that for quite some time.  The
conversion to all lower case was an upstream change that I've just
followed accordingly.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>