systemd: Allow privelged container to have a RW /sys
Systemd has special checks in it for /sys to be mounted RO and then
properly works in a non-root namespace.  This patch performs the check
a different way by looking at /proc/mounts and activating the same
logic when /etc/system-id is a bind mount (which only happens in the
container).

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>