cube-utils: fix cross-env execution
systemd-virt-detect is not always available in a cross-env, so test
for it, and skip if not found.

Also the checks for am_i_a_container were inverted in two cases, so
we fix them.

Now that the am_i_a_container is fixed, we can use that variable to
avoid running container commands in a cross-env.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>