cube-cfg: add 'app' and 'hooks' support
Add the ability to configure the applciation run by the container
as well as the hooks.

The defaults are still the same, but this allows variation on the
options.

Also change all commands to respect the output directory + help
text that indicates the container name is always a valid option.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>