lxc: launch /etc/lxc/lxc-overlayrestore with shell
The behavior of directly launching script is different with launching it
with shell in the case of IMA verification, e.g,
/etc/lxc/lxc-overlayrestore vs sh -c '/etc/lxc/lxc-overlayrestore'. The
formal will trap into execve() syscall where IMA hook just checkes the
integrity of script, but the later doesn't trigger the IMA hook.

If the content of lxc-overlayrestore is always kept, nothing needs to
concern. However, lxc-overlayrestore may be touched by lxc-overlayclean
and thus break its integrity. In order to work around this issue, make
sure lxc-overlayrestore is always called with shell.

Signed-off-by: Lans Zhang <jia.zhang@windriver.com>