images: allow to specify the packages to be not installed
A template may employ IMAGE_INSTALL to install the packages, but
actually not all images really require them. In order to handle this
situation, we have to define CUBE_*_EXTRA_INSTALL in each board-specific
init script to specify the required packages and avoid using IMAGE_INSTALL.

This works well but it is bad for user experience. People knows how to
edit init script to add the necessary layers and use the require directive
to include the required templates because they are part of bitbake syntax,
but people obviously lacks a profound understanding of using
CUBE_*_EXTRA_INSTALL.

In order to resolve this issue, the following actions are made.

- introduce CUBE_*_EXTRA_UNINSTALL.
- re-use IMAGE_INSTALL in template.
- don't use CUBE_*_EXTRA_INSTALL in init script.
- use CUBE_*_EXTRA_UNINSTALL in the private place to blacklist the
  unnecessary packages for certain image.

Signed-off-by: Lans Zhang <jia.zhang@windriver.com>