go: reflect go-cross renamed to go-cross-${TARGET_ARCH}
In meta-virtualization we have renamed go-cross to
go-cross-${TARGET_ARCH} to allow per recipes sysroot to function
correctly and to reflect the fact that this crosstool has arch
specific bindings. We need to carry that rename through to the recipes
in this layer to prevent looking for and not finding go-cross.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>