core: add oci-networking support, docker import
Updating the core system and utilities to allow pflask and oci containers
to attach to the OVS bridge for networking purposes.

Note: only one OCI container can currently be networked, since dhcp is not
present on the interface.

This also adds the ability to add hooks to pflask and OCI containers to
facilitate the network bring up.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>