grub: make it x86/x86_64 specific (take 2)
If you look in history, you'll see a two file implementation of this
which I did because I couldn't get arch overrides to work.  Then
Jason made a cleaner one-file implementation, after testing with
MACHINE based overrides.  However the TARGET_ARCH based overrides
still failed to work, and hence Bruce reverted all related commits
(rightly so) when his x86 image was missing grub.

Re-investigating, and with help from Richard Purdie via IRC, we
now know that packagegroups are special and will default to allarch
if you don't explicitly force something else in advance.

This looks more complicated than it should be, because of the above.
The recommended way to not have this happen is to do the setting
before using the packagegroup.  In fact, if you try to do it _after_
you will trigger a hard error.

In overc, we have overc-common-pkgdefs.inc --- so for all the ones
that source that include, we make the setting there as well.

Cc: Jason Wessel <jason.wessel@windriver.com>
Cc: Bruce Ashfield <bruce.ashfield@windriver.com>
Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>