netcat-openbsd: Unblacklist while waiting on upstream merge of fix.
This was blacklisted because of an implicit dependency on pkg-config.
It largely went unnoticed since most build hosts have such a native
binary present in /usr/bin.  The fix (inherit pkgconfig) was sent
upstream several weeks ago, and is in master-next but not master.

To unblock automated builds of master, lets just do a temporary
bbappend that unblacklists it, while waiting on the upstream merge.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>