avahi: drop minor version skew workaround.
This was added in commit 832fa3457cd98b8918e35793af10bbc1fab5c171i
but now it fails with INC_PR being detected as an unexpanded variable,
since that is no more.

Delete the workaround, which lets us build again, and if the issue
re-appears intermittently then we can revisit it again.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>