mutt: clobber /etc/mime.types to resolve conflict with mime-support
To fix:

   Error: Transaction check error:
     file /etc/mime.types conflicts between attempted installs of
     mutt-1.5.24-r0.core2_64 and mime-support-3.48-r0.core2_64

   Error Summary
   -------------
   ERROR: cube-dom1-0.1-r0 do_rootfs: Function failed: do_rootfs

...we clobber the mutt version of the file and depend on the pkg
mime-support to always be present instead.

We could have done a more complex solution with alternatives being
added to both packages, but since mutt only exists in the legacy
openembedded repo and not the currently maintained repo (hence the
recipe here) -- we simply explicitly choose the version of the file
from the maintained package.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>