recipes-kernel: don't set PREEMPT_NONE
This dates back to when I 1st created the builder, when the
primary objective was a text mode server install that could
build itself.  While we still support that, we've also since
added and achieved new use cases for where this might not be
the right choice anymore.  So remove it and take the default.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>