vim: replace install time sed of defaults with patching
The sed method was fragile and could silently fail to do its
job.  Replace it with a patch to the defaults, so if something
changes or gets renamed etc. we will get a patch failure that
tells us to go check things out.

I'd rather not have to use the patchdir= hack, but that is already
what the main vim.bb does, and fixing vim.bb to have a proper "S"
that would avoid using patchdir appears to be non-trivial.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>