recipes-core: delete tcf-agent from all package lists.
The Target Communication Framework (TCF) package should not be in
package lists, where it can be inadvertently be overlooked and hence
installed and enabled by default.  This is the case for overc:

  commit 1b868bbd0b2a81cd6d09cf2fa8edc41fb56e042a
  Author: Saul Wold <sgw@linux.intel.com>
  Date:   Thu Nov 3 13:11:41 2011 -0700

    Self Hosted Image: Added Task and Images

..as we started from the package lists that trace back to the
above commit, even though we never really needed or wanted the
tcf-agent installed and deployed.

In addition to increasing the attack surface of the machine where
tcf-agent is deployed, I recently noticed that it is also sensitive
to kernel configuration settings ; a custom kernel I was testing
caused tcf-agent init scripts to fail and that triggered the hateful
systemd pointless 1m30s wait during startup/shutdown.

We also fix an obvious tab vs. space issue in one of the files.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>