console: fix "is not a tty" issues
The root cause of the issue is this:

   https://github.com/moby/moby/issues/11462

Since /dev/pts is mounted with newinstance the stdin/stdout of the console
you in when you invoke 'cube-console' is not visible to the end point (since
its /dev/pts is unique). agetty and other commands try to open /proc/self/fd/0,1,2
to determine the tty stats .. and of course they are invalid and the commands
fail with "unable to determine tty name" (or a variant).

While this issue is fixed in newer runc variants, we can't update right now
due to the scope of the changes in the console handling in those new variants
(but when docker updates, we'll update).

BUT. There is a workaround. If you wrap the command in 'script', the tty type
can be determined by the inside application, and everything works like you
expect.

So we add yet another layer of nesting around our console implementation and
invoke agetty inside script (and dtach!!). The result is a series of working
consoles :D

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>