plfask:  Fix parallel make problem
Two copies of gcc can get running on the same c file to the same
location and you get an error like:

tmp/work/core2-64-wrs-linux/pflask/git-r0/git/build/src/sync.1.o: file not recognized: File truncated

The install rule depends and builds all the objects so the build rule
is effectively not used.  With the build rule there it will run in
parallel to the install objects rules which build the same files.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>