Housekeeping: Set S if recipe has a SRC_URI but doesn't use default S
Recipes which have a SRC_URI but don't unpack/fetch to the default S
directory should set 'S = "{WORKDIR}"'. Using the default S in these
cases used to cause a QA Warning but a bug which is currently always
creating the default ${WORKDIR}/${BPN}-${PV} directory results in the
QA Warning being 'hidden'.

For folks using meta-overc with older Yocto releases this will cut
down on the build noise. Since the bug which was 'hiding' the QA
warning has been raised with the associated devs by fixing this now we
avoid seeing the QA warnings when/if the bug is fixed.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>