cube*: enable VT h/w processing
This commit restores the ability to pass VT devices to containers so
that systemd logind can execute against a dedicated VT.

To do this, we add the ability to bind mount /dev/tty devices to
containers, and tweak the systemd configuration to use any devices
that are available.

This configuration is dynamic at runtime and will respect cube-cfg
changes after a container is installed.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>