cube-cmd-server: enhance for extensibility and authentication
The existing cube-cmd-server was written to prove that IPC was possible
between containers, and that a single control container could handle
requests on behalf of the larger system.

The implementation used a number of schemes, nanomsg, dtach, ACLs, etc,
which were all proven to work. That being said, to allow integration
with orchestration systems such as kubernetes a more extentible system
was required, where the commands, clients and authentication is all
dynamic.

With this change, we move to a nanomsg only structure, where all
control and commands are run over nanomsg channels, and the ability
for finer grained error handling is possible.

The new structure runs over a modified nanocat which handles the
transmission of stdin/stdout over the channels between the client
and server.

With that change we now have dynamic ACLs, the ability to add
new commands on the fly and simpler set of control and command channels
to maintain. Any container can be added to the system, without the
need to provision each one for monitoring.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>