networking: reliability fixups and dynamic netprime
This commit brings reliability to the startup of container networking
and a new feature of dynamic netprimes.

It also allows the stacking and redirection of local ports for integration
with orchestration systems.

The significant change here is the fixup of the hooks to manage/control
dhclient integration from essential -> container namespaces. With a custom
dhclient-script, we can now dynamically request IPs without overwriting
essential's resolv.conf (which is sort of important!).

The command 'cube-ctl netprime <container name>' is now functional and
allows the runtime switching of the container responsible for networking
traffic.

As part of that switch, we do not use any of the container's resources
to configure traffic, since we cannot put dependencies on the contents
of the container.

The result is that dns/dhcp is move to a 'vrf' container, which is
currently dom0. But all dns/dhcp still points to the netprime, which
has rules configured to forward the traffic to the VRF.  All containers
in the system are configured to get DNS from the netprime, which
transparently subcontracts it to a container that we can be sure has
the right capabilities and configuration to provide the service.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>