python-ansible: uprev to v2.3.1.0 to allow for python3 support
The upstream ansible project has/is working on ensuring ansible works
with python3 but this support is only available after v2.2.0.0. In
order to allow us to make a python3 version of ansible available I am
therefor completing this uprev first.

This was tested with a local ping test, a minimal playbook and the
default cube-essential playbook, all work as they did previously.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>