packagegroups: sync to meta-oe/oe-core june 2017
Updating the bbappends and package lists to work with the oe-core/meta-oe
from June 6 2017.

A fairly normal uprev with the except of dhcp-libs, which should be automatically
picked up as a dependency and installed via the packagegroup. But the DNF resolver
is missing this depdendency and we end up with a non functional dhclient due to
missing shared libraries.

To fix that, we temporarily add an explicit 'dhcp-libs' to the image install
of our images and we are good to go.

At the same time, we bump the PVs of the images, since they are due to be rev'd.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>