wiggle: disable new gcc7 case statement fall through warning
With the move to gcc v7 we now get lots of these:

   ccan/hash/hash.c:651:5: error: this statement may fall through [-Werror=implicit-fallthrough=]

Since the authors chose to use quite a few of these in their source,
we simply disable it for the whole package.

In the process we also found that the CFLAGS weren't being used to
build the package, so we fix that at the same time.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>