cube-cmd-server: Make startup deterministic
Other services depend on the cube-cmd-server being ready to serve
requests. On a system which has many startup tasks occasionally the
cube-cmd-server would start but not be ready to serve a command.  This
would cause dom0 not to start properly or dom0 would not be able to
startup other peer containers.

With the systemd-notify the other processes that depend on the
cube-cmd-server will be properly gated until it is started.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>