cube-cfg: Add detached console support for runc create
The console socket needs to be set to /dev/null as well as to have a
runc that is patched to turn off the console terminal without changing
the config.json.

This patch is a step along the way to having a console socket that is
available at runtime dynamically.  For now the console is turned off
providing the same capability that existed in OverC previously.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>