overc-conftools: we can't check just for tzinfo
Per the ansible documentation "If a task fails or is skipped, the
variable still is registered with a failure or skipped status, the
only way to avoid registering a variable is using tags.". So the check
for tzinfo will always succeed and is insufficient. Instead we check
for tzinfo.stat.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>