dtach: Add console socket option and pty link location.
The console socket option will open listening socket which will work
with runc which passes a file descriptor to use as the pty.

The pty link location will open an additional pty which can
be connected directly by another application.   The idea is to
provide a pty redirector for a run console.  The pty
link can be shared to a different name space by using a bind
mount.

This allows for a small compact console connection manager.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>