console: only attach to screen-tty if screen-tty is the console manager
We were previously attaching to screen if 'tty' was the console manager.
Since tty was the default, this meant that all containers (app, system,
etc) would attach to screen and display nothing.

We really only want to attach to screen if screen-tty is the manager,
or if a container is a system container (i.e. is running 'init').

By setting 'console' as a property in the install configuration, screen-tty
will be set, hence we can control it for system containers.

We can also inspect the entry point, and if it is 'init*' set screen-tty
as the manager.

Finally, the setting can always be changed via cube-cfg + cube-cfg gen
at a later point.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>