overc-system-agent: replace dom0_contctl calls
commit ea8e4428f49f93c1afd49f3168a0b52b24e953f3 [cube-cmd-server:
enhance for extensibility and authentication] dropped the dom0_contctl
script from existence, but it remained in use in the 'dom0' script,
breaking its use. Swap out calls to dom0_contctl for calls to the new
'c3' tool. For the most part the behavior is restored but differences,
for example with the values returned with '-L' (list), could mean that
callers will need some tweaking.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>