dom0: esssential: use shared linux firmware
Currently we ship the linux-firmware in -essential this offers some
direct and indirect challenges.

The direct challenges are that the size of -essentially is almost
double what it can be without the firmware include, additionally the #
of reasons to require an update to -essential is greater with the
firmware included.

An indirect challenge is that dom0 will have to interact with
-essential's rootfs should a container wish to add kernel firmware to
the system.

By using the 'bridge' filesystem (/var/lib/cube/essential) and
shipping the firmware in dom0 we can offer solutions to the above
challenges. We remove the added size of firmware in the -essential
fileystem (and images), we affect firmware updates by updating dom0
and dom0 can contribute to firmware by simply manipulting its own
filesystem.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>