dom0-init: rename dom0-contctl to dom0-ctl-core
We never really liked the name and since we are changing many things
in preparation for the new 'oci' container management approach we are
taking the opportunity to rename this now. The new name is inline with
'cube-ctl' and we append a '-core' string as we might have future
additions such as 'dom0-ctl-disk' or similar. This handles both
startup and shutdown so using 'init' or 'startup' wasn't a good fit.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>