meta-cube: refresh vimrc adjustment patch
Apparently vim learned to not jump to last location on commit
logs, which was one of the most annoying problems with the
feature.  In making that change, it breaks our vimrc patch.

Drop the chunk where we patched out that setting and we'll give
it a whirl and see if it is still annoying on any other files.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>