pflask: Add the ability to use a wildcard network interface spec
It is desirable to allow pflask to work with what ever hardware is
detected as well as specific hardware.  The --netif=Interface was
extended with this patch to allow a "+" which will match any other
interface with the same prefix.  For example --netif=enp+:enp+ will
select all available interfaces that start with enp like enp0s3,
enp0s4, etc...

Also clean up instances of NETWORK_DEVICE_CLASSES which is no longer used.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>