tools: introduce ostree
The latest skopeo needs ostree to work with this type of packed directory
structure.

There are two versions of ostree in different layers: meta-updater and
meta-flatpak (in intel-iot-refkit), but both resulted in packaging
errors due to our multilib configuration.

While those different upstream options are fixed, we take a copy of
the meta-flatpak variant and fix the issues.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>