dom0-init: stop cube containers before shutting down
It was observed that the system hang for up to 1Min 30Secs when users
do rebooting/shutting down on dom0 or cube containers, with the
following messages showing on the console:
|
| [ **] A Stop job is running in container dom0...

this is caused by the cube containers started by dom0 are still running
during system shutting down, and we are missing a stop process in
dom0-containers.

Fixed by implementing a stop process in dom0-containers, and also,
dom0-ctl-core's ExecStop program must run before
container-shutdown-notifier, or else it's not going to succeed to stop
the containers during cube-essential shutting down procedure(triggered
by container-shutdown-notifier).

Signed-off-by: Ming Liu <ming.liu@windriver.com>