meta: drop explicit dhcp-libs references
dhcp-libs would be set RDEPENDS to the packages that requiring it, like
dhcp-client, dhcp-server and so on, which is done in package_do_shlibs
during do_package task, this is a common procedure processed by bitbake
to all the recipes, if there is anything wrong in it, we should fix that
instead of hiding it with this temporary workaround, this sort of hack
looks redundant and will cause mantaining cost.

I had verified the dhcp-libs package installed just fine as expected
with this change, so I attempt to say we probably are safe to drop the
workarounds now.

Signed-off-by: Ming Liu <ming.liu@windriver.com>