systemd: Work around udev requiring a devtmpfs mount
When using containers the /dev directory is a private tmpfs by
default.  The udev rules have an implicit assumption that /dev is
mounted as a devtmpfs where the kernel auto instantiates a device node
prior to issuing a uevent notification via the sysfs.

If you turn on logging for systemd-udev as follows you can see the problem:

echo 'udev_log="debug"' >> /etc/udev/udev.conf
systemctl restart systemd-udevd

Now in the system journal if you have a USB hotplug even it the
following error will show up:

Oct 19 10:05:18 localhost systemd-udevd[535]: starting 'ata_id --export /dev/sda'
Oct 19 10:05:18 localhost ata_id[535]: unable to open '/dev/sda'
Oct 19 10:05:18 localhost systemd-udevd[534]: Process 'ata_id --export /dev/sda' failed with exit code 1.

This is because the udev rules assume /dev/sda was auto instantiated by the devtmpfs.

The solution is to check for, and create if missing the device node in
the default rule template via an import{program} rule.  This will run
immediately prior to running any of the rule templates that have the
implicit assumptions.  No changes are needed to the hot plug services
because udev will issue the rest of the events normally so long as the
device node exists initially.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>